requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
argon2-cffi==23.1.0
SQLAlchemy==2.0.23
alembic==1.12.1
PyJWT==2.8.0
email-validator==2.1.0.post1
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
